feat: 글 작성 완료 후 모달이 커뮤니티에서 뜨도록 구현
